refactor: be - session.util 이동 후 renanme #66
