+++ backend/src/main.ts
@@ -1,6 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import { createSessionMiddleware } from './util/session.util';
+import { createSessionMiddleware } from './middlewares/session.middleware';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
